refactor: srcs int src dir, import path fix, delete unused libraries
