+++ src/nodes/base_node.py
