+++ src/nodes/execute_sql.py
@@ -1,8 +1,8 @@
 import pandas as pd
 import re
-from models.graph_state import GraphState
-from utils.data_utils import filter_csv_with_sql, format_dataframe
-from nodes.base_node import BaseNode
+from ..models.graph_state import GraphState
+from ..utils.data_utils import filter_csv_with_sql, format_dataframe
+from .base_node import BaseNode
 
 
 class ExecuteSQLNode(BaseNode):
