+++ src/nodes/retrieve_from_web.py
@@ -1,10 +1,10 @@
-from models.graph_state import GraphState
+from ..models.graph_state import GraphState
 from configs.default_translate_params import DefaultTranslateParams
-from utils.translate import ko_to_eng
+from ..utils.translate import ko_to_eng
 from configs.default_web_params import DefaultWebParams
-from utils.web_search import web_search
+from ..utils.web_search import web_search
 
-from nodes.base_node import BaseNode
+from .base_node import BaseNode
 
 
 class WebSearchNode(BaseNode):
